+++ example/bug-fixed/depend-shared-kind-target-run/src/main.cc
@@ -1,0 +1,8 @@
+#include "add.h"
+#include <iostream>
+
+int main() {
+    int32_t result = add(1, 2);
+    std::cout << "result: " << result << std::endl;
+    return 0;
+}