+++ example/bug-fixed/depend-shared-kind-target-run/test.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+xmake build -r
+
+xmake run main